fix(SPV3): Only clear SPV3's log if it's larger than 1 MiB

+++ src/Main.cs
@@ -1,15 +1,15 @@
 /**
  * Copyright (c) 2019 Emilian Roman
  * Copyright (c) 2020 Noah Sherwin
- * 
+ *
  * This software is provided 'as-is', without any express or implied
  * warranty. In no event will the authors be held liable for any damages
  * arising from the use of this software.
- * 
+ *
  * Permission is granted to anyone to use this software for any purpose,
  * including commercial applications, and to alter it and redistribute it
  * freely, subject to the following restrictions:
- * 
+ *
  * 1. The origin of this software must not be misrepresented; you must not
  *    claim that you wrote the original software. If you use this software
  *    in a product, an acknowledgment in the product documentation would be
@@ -42,11 +42,12 @@ namespace SPV3
     {
       Directory.CreateDirectory(Paths.Directory);     /* create data directory */
 
-      WriteAllText(Paths.Exception, ""); // ...clear log.
+      if (Exists(Paths.Exception) && new FileInfo(Paths.Exception).Length > 0x100000) // If larger than 1 MiB, ...
+        WriteAllText(Paths.Exception, ""); // ...clear log.
 
       /**
        * We determine installation or initiation mode:
-       * 
+       *
        * -   initiation: The HCE executable exists, thus SPV3 is ready to be loaded.
        * -   installation: The manifest exists, thus SPV3 is ready to be installed.
        *
